+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Juniper.jl Changelog
 
+## Unreleased
+- Bugfix: No bounds on binary variables in Feasibility pump [#143](https://github.com/lanl-ansi/Juniper.jl/issues/130)
+
 ### v0.4.2
 - Ability to not accept `almost` MOI solver status codes in tree search
 - Better handling of MOI solver status codes internally